feat: modify CTA button + add sticky picture in desktop from problem solution section
