+++ app/globals.css
@@ -22,7 +22,9 @@
     --popover: 0 0% 100%;
     --popover-foreground: 222.2 84% 4.9%;
 
-    --primary: #010658;
+    /* --primary: #010658; */
+    --cta-button: var(--palette-black);
+    --primary: var(--palette-blue);
     --primary-hover: #04127c;
     --primary-foreground: 210 40% 98%;
