+++ components/ui/nav.tsx
@@ -5,20 +5,7 @@ import Link from "next/link";
 import { MAIL_TO } from "@/hooks/useMailTo";
 import React from "react";
 import MobileNav from "./mobile-nav";
-
-type NavLinkProps = React.PropsWithChildren<React.ComponentProps<typeof Link>> & {
-  href: string;
-};
-
-const NavLink = ({ children, href, ...props }: NavLinkProps) => {
-  return <li className="relative group list-none">
-    <Link href={href} {...props}>
-      {children}
-    </Link>
-    <span className="absolute -bottom-2 left-1/2 w-0 h-0.5 bg-[#E7F3FF] group-hover:w-1/2 group-hover:transition-all"></span>
-    <span className="absolute -bottom-2 right-1/2 w-0 h-0.5 bg-[#E7F3FF] group-hover:w-1/2 group-hover:transition-all"></span>
-  </li>
-}
+import NavLink from "./nav-link";
 
 const Nav = () => {
   return (
