+++ tailwind.config.ts
@@ -24,6 +24,7 @@ const config = {
         ring: "hsl(var(--ring))",
         background: "hsl(var(--background))",
         foreground: "hsl(var(--foreground))",
+        cta: "var(--cta-button)",
         primary: {
           DEFAULT: "var(--primary)",
           hover: "var(--primary-hover)",
